chore: extract presence suite into src/suites/presence.ts
Part of REAL-1037.
